// src/app/models/barber.ts
export interface Barber {
	id: string
	name: string
	email: string
	image: string
	specialties: string[]
	availability: Availability
	bio: string
	contactNumber: string
}

export interface Availability {
	monday: string
	tuesday: string
	wednesday: string
	thursday: string
	friday: string
	saturday: string
	sunday: string
}

export interface BarberService {
	barberId: string
	services: Service[]
}

export interface Service {
	id: string
	name: string
	price: number
	duration: string
	description: string
}